[Wasm] Adapt utilities to 64-bit pointers.
Adapts storage utilities like `Buffer`s and `HashTable`s
to be able to hold up to 2^64 values. Therefore, also use
64-bit tuple IDs.

+++ unittest/backend/WasmTestV8.cpp
@@ -144,7 +144,7 @@ struct invoke_v8<PrimitiveExpr<ReturnType, ReturnL>(PrimitiveExpr<ParamTypes, Pa
         env->Set(context, mkstr(*isolate_, "insist"), func_insist).Check();
         auto func_throw = v8::Function::New(context, _throw).ToLocalChecked();
         env->Set(context, mkstr(*isolate_, "throw"), func_throw).Check();
-        Module::Get().emit_function_import<void(void*,uint32_t)>("read_result_set");
+        Module::Get().emit_function_import<void(void*,uint64_t)>("read_result_set");
         auto func_read_result_set = v8::Function::New(context, read_result_set).ToLocalChecked();
         env->Set(context, mkstr(*isolate_, "read_result_set"), func_read_result_set).Check();
         Module::Get().emit_import<uint64_t>("alloc_addr_init");
